+++ utils/di/index.ts
@@ -1,2 +1,3 @@
-export { Container, createToken } from './container';
+export { Container } from './container';
+export { createToken } from './modules';
 export type { Lifetime, Token, TypeMap } from './type';
